+++ src/Libraries/CG.Orange.Abstractions/Managers/IProviderPropertyManager.cs
@@ -49,7 +49,7 @@ public interface IProviderPropertyManager
     /// number of <see cref="ProviderPropertyModel"/> objects in the underlying storage.</returns>
     /// <exception cref="ManagerException">This exception is thrown whenever the
     /// manager fails to complete the operation.</exception>
-    Task<long> CountAsync(
+    Task<int> CountAsync(
         CancellationToken cancellationToken = default
         );
